v1.9.54
Assassin flavor added
Kill Quests added in fountain (max of 3 per person)
Fountain quests have a rarity system that details how many you need to kill, how much you get for it, and if you get an item that the enemy drops. If you have the needed rarity it is a 100% guaranteed drop. However, it might just become a tome (elite creeps)